Add link to "Wiki" in the readme file
There are 2 reasons for this change:

- Reading from mobile, I could not see a wiki tab in the navigation.
- Since the expected audience might include newbies, it might be helpful to just give them the link

The 2nd reason is why I added the link to the beginning of the file as well. It's also useful to get to the repository, and just "Click the big link" to get to the FAQs, even though we'll likely be sharing the wiki link more often than the project root.